Wave 13: 47 functions matched, 1210 -> 1257, and the large band is a step not a slope
Eight agents on the >=256-byte straight-line band. All 47 verified with trymatch
before promotion; both builds still reproduce SHA1
14dd0b22c894865867aff89e8116b2dffae25605, which also regression-proves the 1210
already-promoted functions. 21,744 bytes, byte progress 5.65% -> 9.48%.

Targeting was `backward_branches == 0, size >= 256` -- 249 functions, 121,484
bytes, 21% of all remaining ROM code. Bands closed: 6 at >=768B, 10 at 512-768,
12 at 384-512, 19 at 256-384.

COST PER BYTE IS FLAT TO 512, THEN STEPS. Measured per band, attempts spent on
functions that closed: 258 (wave 12, 280-368), 278 (384-512), 438 and 574 (two
agents, 512-768), 811 (768+). 278 vs 258 is inside the noise of a six-function
sample, so the 368/384 boundary is not a discontinuity -- 280-512 is one regime.
The planning question is which side of the step a target is on, not how big it
is. One agent's first figure of 350 circulated for an hour before it corrected
its own arithmetic; the number here is the corrected one.

What actually predicts cost is NEW AGGREGATE TYPES, not size. Three agents
converged on it independently: the one-attempt matches needed zero new structs,
and the target abandoned as too expensive is 400 bytes -- smaller than three
matches -- and expensive only for three new ROM tables. Sharpened by the
512-768 pair: new *symbols* do not predict, only new *layouts* do (a 624-byte
function needed 24 new symbols, all scalars or flat arrays, and matched first
try). The 512-768 agent dissents with receipts and puts the crossover above
768, not inside its band. Wave 14 should sort by distinct new struct layouts.

THE `.rodata` POOL IS THE REAL BLOCKER, AND IT IS NOT LOOP-TRIGGERED. The doc
said -fforce-addr parks an address in .rodata when it is live across a LOOP,
and offered the family's two loopless members as the control that pinned it.
Four agents refuted it independently: a controlled probe (same source, 1 call vs
19), a merge-point probe, and two existence proofs on straight-line functions
that carry a pool word anyway. The trigger is reference count across a
control-flow merge. Corrected in the doc; the surviving part of the old claim is
that a PAIR of words is loop-driven while a single word is not.

A cheap screen for it, proposed by one agent and refined by another after it
flagged already-promoted work: fan-in == 1 AND run-density >= 4 on the 4-byte
grid. A .LC pool is consecutive words; real single-use ROM data is isolated.
Controls separate 0-2 against 9-16. Over the band it flags 66 functions, 34,432
bytes. NEITHER HALF IS SUFFICIENT: fan-in alone would park the proc-script
wrappers that are the cheapest family in the tree, and density alone frees a
function that is genuinely blocked. It is a prioritisation heuristic, never a
verdict.

And diagnose it ROM-side, not candidate-side. A draft carrying more live values
than the original emits its own .rodata pool on a function whose ROM is clean --
that is register pressure in your C, fixable, not a park. I had broadcast the
opposite and it would have parked matchable functions.

THE BLOCKER IS ALSO PARTLY SOLVED. One agent produced a byte-exact AND
relocation-exact match through a .LC word by binding the pool word's ADDRESS to
a local rather than dereferencing it in place -- the two existing sections were
right that in-place fails and wrong that the workaround therefore fails. And the
seven parked at 99.5% are decidable per-function after all: the candidate's
.rodata is fully symbolic through .rel.rodata, so its value can be compared
against baserom.gba. parked.json's "no per-function check can see it" is false.
Verification and placement were conflated; only placement is linker work.

Three harness bugs, each of which had been silently costing matches:
- `awlib.BRANCH_TARGET_RE` omitted `bhs`/`blo`, the alias spellings objdump
  emits. 16 functions the index called straight-line have a back edge; four are
  >=256B and one was handed out this wave with two nested loops.
- The permuter aborted at load on any source reaching hardware.h, because
  pycparser cannot parse `__attribute__((aligned(4)))`. That removed every
  hardware-touching function from it. Fixed after measuring that stripping the
  attribute is byte-neutral on seven promoted files.
- `index_functions.decompiled_symbols()` missed signatures whose parameter list
  does not close on the first line -- the same failure already documented for
  function-pointer returns. sub_08036F68 was in the ROM, with both builds green,
  while the index called it `asm` and the queue would have reissued it.

Two reconciliations the per-function check cannot see, because trymatch compiles
one unit. Eight prototypes collided with promoted definitions at link time. Seven
resolved to the definition; the eighth resolved the other way and is the
instructive one: `s16 sub_080674F4` is byte-identical in its own body (the table
is `const s16[]`, so the ldrsh sign-extends anyway) but forces callers to
re-narrow -- +4 bytes at each of three call sites, +12 across the ROM, which is
exactly the size delta that broke the first build. Settled from the callers, as
the "stub's return type carries no information" rule requires.

Also corrected: a high-register prologue does NOT mean a loop. Two agents found
it independently on loop-free functions -- one holding values live across calls,
one holding CSEd bitfield masks. It means five-plus simultaneously live values;
identify WHAT is live before concluding anything about control flow.

+++ src/decomp/c_08034F6C.c
@@ -7,7 +7,11 @@
  * sub_08034F6C @ 0x08034F6C, sub_08034F7C @ 0x08034F7C
  */
 
-s8 sub_08034F6C(void)
+/* `int` and not `s8`: byte-identical here either way (the `lsls #0x18;
+ * asrs #0x18` is the s8 member read), but sub_080345C8 tests the result with a
+ * bare `cmp r0, #0` and an s8 return would have narrowed it there first.
+ * Retyped in wave 13 (A2); re-verified byte-for-byte. */
+int sub_08034F6C(void)
 {
     return gUnknown_030030F0.unk00;
 }
